move get_entity* helper to test
